+++ src/pages/product-detail.js
@@ -1,0 +1,40 @@
+import { Footer } from "../components/Footer";
+import { Header } from "../components/Header";
+import { Layouy } from "../components/Layout";
+import { Component } from "../core/Component";
+import { router } from "../core/router";
+import { html } from "../utils/html";
+
+export function ProductDetailPage($root) {
+  const layout = new Layouy({
+    header: new Header({ nav: new Nav() }),
+    main: new Main({}),
+    footer: new Footer(),
+  });
+  document.querySelector($root).innerHTML = html`${layout}`;
+  layout.setup();
+  console.log(router.getParams());
+
+  return () => {
+    layout.dispose();
+  };
+}
+
+class Nav extends Component {
+  renderContainer() {
+    return html` <div ${this.dataAttribute.attribute} class="flex items-center space-x-3">
+      <button onclick="window.history.back()" class="p-2 text-gray-700 hover:text-gray-900 transition-colors">
+        <svg class="w-6 h-6" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+          <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M15 19l-7-7 7-7"></path>
+        </svg>
+      </button>
+      <h1 class="text-lg font-bold text-gray-900">상품 상세</h1>
+    </div>`;
+  }
+}
+
+class Main extends Component {
+  renderContainer() {
+    return html`<main ${this.dataAttribute.attribute} class="max-w-md mx-auto px-4 py-4"></main>`;
+  }
+}
